requirements.txt
discord.py==1.4.1
configparser==5.0.0
moviepy==1.0.3
pytube3==9.6.4
Pillow==7.2.0